requirements.txt
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
redis
gunicorn
mysql-connector